catch error

+++ app/Http/Controllers/RPController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Roles;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class RPController extends Controller
 {
@@ -18,6 +19,17 @@ class RPController extends Controller
     }
 
     public function rpCreate(Request $req) {
+        $validator = Validator::make($req->all(), [
+            'name' => 'required',
+            'permission_level' => 'required|integer'
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json([
+                'errors' => $validator->errors()->first()
+            ], 403);
+        }
+
         $role = Roles::create([
             'name' => $req->name,
             'permission_level' => $req->perm
@@ -29,6 +41,18 @@ class RPController extends Controller
     }
 
     public function rpPatch(Request $req) {
+        $validator = Validator::make($req->all(), [
+            'id' => 'required|integer',
+            'name' => 'required',
+            'permission_level' => 'required|integer'
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json([
+                'errors' => $validator->errors()->first()
+            ], 403);
+        }
+
         $role = Roles::findOrFail($req->id);
 
         $role->update([
@@ -42,6 +66,16 @@ class RPController extends Controller
     }
 
     public function rpDelete(Request $req) {
+        $validator = Validator::make($req->all(), [
+            'id' => 'required|integer',
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json([
+                'errors' => $validator->errors()->first()
+            ], 403);
+        }
+
         $role = Roles::findOrFail($req->id);
         $role->delete();
